+++ plugins/generic-sessions/protocol_lws_messageboard.c
@@ -330,7 +330,7 @@ callback_messageboard(struct lws *wsi, enum lws_callback_reasons reason,
 			return -1;
 		n = lws_write(wsi, start, p - start, LWS_WRITE_HTTP_HEADERS);
 		if (n != (p - start)) {
-			lwsl_err("_write returned %d from %d\n", n, (p - start));
+			lwsl_err("_write returned %d from %ld\n", n, (long)(p - start));
 			return -1;
 		}
 		s[0] = '0';
